+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 <a name="unreleased"></a>
 ## [Unreleased]
 
+
+<a name="v3.1.3"></a>
+## [v3.1.3] - 2022-02-24
 ### Features
 - **fidget:** moved to fidget nvim from lsp-status
 - **golang:** added go test workspace
@@ -657,7 +660,8 @@
 - initial commit
 
 
-[Unreleased]: https://github.com/tigorlazuardi/nvim/compare/v3.1.2...HEAD
+[Unreleased]: https://github.com/tigorlazuardi/nvim/compare/v3.1.3...HEAD
+[v3.1.3]: https://github.com/tigorlazuardi/nvim/compare/v3.1.2...v3.1.3
 [v3.1.2]: https://github.com/tigorlazuardi/nvim/compare/v3.1.1...v3.1.2
 [v3.1.1]: https://github.com/tigorlazuardi/nvim/compare/v3.1.0...v3.1.1
 [v3.1.0]: https://github.com/tigorlazuardi/nvim/compare/v3.0.3...v3.1.0